Extract out connection monitor mod
